Fixed errors

+++ app/src/main/java/com/example/imran_mamirov_hw_7_6/ui/fragment/taskList/TaskListViewModel.kt
@@ -20,7 +20,7 @@ class TaskListViewModel(
         }
     }
 
-    fun deleteTask(taskId: Int) {
+    fun deleteTask(taskId: Long) {
         viewModelScope.launch {
             deleteTaskUseCase(taskId)
         }
