+++ domain/src/main/java/com/example/domain/interfaces/TaskRepository.kt
@@ -7,6 +7,6 @@ import kotlinx.coroutines.flow.Flow
 interface TaskRepository {
     suspend fun insertTask(taskEntityModel: TaskEntityModel)
     suspend fun getTaskById(taskId: Int): TaskEntityModel?
-    suspend fun deleteTask(taskId: Int)
+    suspend fun deleteTask(taskId: Long)
     suspend fun fetchTask() : Flow<List<TaskEntityModel>>
 }
