+++ domain/src/main/java/com/example/domain/model/Task.kt
@@ -1,8 +1,8 @@
 package com.example.domain.model
 
 data class Task(
-    val taskId: Int = 0,
+    val taskId: Long = 0,
     val taskName: String,
     val description: String,
-    val time: Int
+    val time: Long
 )
