+++ domain/src/main/java/com/example/domain/usescase/DeleteTaskUseCase.kt
@@ -3,7 +3,7 @@ package com.example.domain.usescase
 import com.example.domain.interfaces.TaskRepository
 
 class DeleteTaskUseCase(private val taskRepository: TaskRepository) {
-    suspend operator fun invoke(taskId: Int) {
+    suspend operator fun invoke(taskId: Long) {
         taskRepository.deleteTask(taskId)
     }
 }
